# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "GPSLink Gateway",
        "description": "Available API for GPSLink Gateway",
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/assignment/create": {
            "post": {
                "tags": [
                    "Assignment"
                ],
                "summary": "Create Vehicle Assignment Info",
                "operationId": "CreateVehicleAssignment",
                "requestBody": {
                    "description": "Vehicle Assignment Information",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "vehicle_id": {
                                        "type": "integer"
                                    },
                                    "driver_name": {
                                        "type": "string"
                                    },
                                    "mileage": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "vehicle_id": 1,
                                    "driver_name": "Juan Dela Cruz",
                                    "mileage": 1825
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Vehicle assignment is successfully registered",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/VehicleAssignment"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "409": {
                        "description": "Vehicle assignment already exist!"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/assignment/assignmentById/{id}": {
            "get": {
                "tags": [
                    "Assignment"
                ],
                "summary": "Get vehicle assignment by id",
                "operationId": "GetVehicleAssignmentById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/VehicleAssignment"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Customer not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/assignment/list": {
            "post": {
                "tags": [
                    "Assignment"
                ],
                "summary": "Get list of Assignments",
                "operationId": "AssignmentList",
                "requestBody": {
                    "description": "Vehicle Id - NOTE: If vehicle_id object is omitted then all vehicle assignments will be return.",
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "vehicle_id": {
                                        "type": "integer"
                                    },
                                    "transporter_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "vehicle_id": 0,
                                    "transporter_id": 0
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "ok"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/assignment/update/{id}": {
            "put": {
                "tags": [
                    "Assignment"
                ],
                "summary": "Updated Assignment",
                "description": "Update Assignment information.",
                "operationId": "UpdateAssignment",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id to be updated",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Updated Vehicle Assignment object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    },
                                    "vehicle_id": {
                                        "type": "integer"
                                    },
                                    "driver_name": {
                                        "type": "string"
                                    },
                                    "mileage": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "id": 1,
                                    "vehicle_id": 1,
                                    "driver_name": "Juan Dela Cruz",
                                    "mileage": 1825
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Vehicle Assignment is successfully updated!"
                    },
                    "404": {
                        "description": "Vehicle Assignment not found"
                    },
                    "409": {
                        "description": "Vehicle Assignment Id does not matched!"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/assignment/delete/{id}": {
            "delete": {
                "tags": [
                    "Assignment"
                ],
                "summary": "Delete Assignment by id",
                "operationId": "DeleteAssignment",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Vehicle assignment is successfully deleted!",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/VehicleAssignment"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Vehicle assignment does not exist!"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "VehicleAssignment": {
                "title": "VehicleAssignments",
                "description": "VehicleAssignments",
                "required": [
                    "vehicle_id",
                    "driver_name",
                    "mileage"
                ],
                "properties": {
                    "id": {
                        "description": "ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "vehicle_id": {
                        "description": "Vehicle ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "driver_name": {
                        "description": "Driver name",
                        "type": "string"
                    },
                    "mileage": {
                        "description": "Mileage",
                        "type": "integer",
                        "format": "int64"
                    },
                    "register_by_user_id": {
                        "description": "Register By UserId",
                        "type": "integer",
                        "format": "int64"
                    },
                    "updated_by_user_id": {
                        "description": "Updated By UserId",
                        "type": "integer",
                        "format": "int64"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost/oauth/authorize",
                        "tokenUrl": "http://localhost/oauth/token",
                        "refreshUrl": "http://localhost/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    }
}